registry/requirements.txt
pyzmq>=25.1.2
msgpack>=1.0.7